requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
Faker==26.0.0
Markdown==3.6
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
validate-docbr==1.10.0
